// webview/action-types.ts
export type ActionKeys =
  | 'Home'
  | 'Login'
  | 'Search'
  | 'Consign'
  | 'ConsignGuide'
  | 'PhoneVerify'
  | 'QuickPurchase'
  | 'CollectBook'
  | 'Wishlist'
  | 'Store'
  | 'GuideBuy'
  | 'Mypage'
  | 'Favorite'
  | 'FAQ'
  | 'ShareLink'
  | 'Pass'
  | 'Twitter'
  | 'Kakao'
  | 'LookUp'
  | 'BuzzBooster'
  | 'BuzzBoosterDefault'
  | 'notificationSetting'
  | 'deviceNotificationSetting'
  | 'PhotoCardDetail'
  | 'Amplitude'
  | 'ConsignHistory'
  | 'KakaoPlus'
  | 'QuickPurchaseFilter'
  | 'PhotoCardFilter'
  | 'ConsignDiscard'
  | 'saveImage'
  | 'Notifly'

export type getActionKeys = 'DeviceNotification'